lambda/code_commit_post_hook/requirements.txt
python-jenkins>=0.4.13
setuptools
